server/requirements.txt
alembic==1.3.1
aniso8601==8.0.0
ansicolors==1.1.8
blinker==1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Deprecated==1.2.6
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
flask-buzz==0.1.14
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-praetorian==1.0.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
passlib==1.7.2
pendulum==2.0.5
poetry==1.0.0
py-buzz==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyGithub==1.43.8
PyJWT==1.7.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
pytzdata==2019.3
rauth==0.7.3
requests==2.22.0
rfc3339==6.2
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
